requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.20
pydantic==2.11.9
scipy==1.16.2

# Audio processing
torch==2.8.0
torchaudio==2.8.0
soundfile==0.13.1
librosa==0.11.0

# Transcription
openai-whisper==20250625
transformers==4.56.1

# Separation
demucs==4.0.1

# Diarization (optional)
pyannote.audio==3.4.0

# Utilities
requests==2.32.5
python-dotenv==1.1.1

# Development tools
pytest==8.4.2
httpx==0.28.1
black==25.1.0
ruff==0.13.0
mypy==1.18.1